Fix page regex matching on numbers in the body of text
Previously the regex matched page titles in the format 'X.X' where X is any numeral.

Unfortunately this broke in very specific cases, like on 11.45 where Bastion mentions a spell load of "37.3" scrolls. 

The new regex looks for page titles in that format that also have one or more newlines immediately after them.

Notably this does mean that if the transcript file is in CLRF instead of LF (basically,  if you're hosting the bot on Windows) it'll break horrifically 'cuz CLRF uses '\r\n' instead of '\n'. It could be nice to account for this in the future by putting some sort of "\n OR \r\n" in there but the regex is already pretty heinous as-is and at the moment this bot only has one deployment that I know of.